requirements.txt
# Core Dependencies
annotated-types==0.7.0
anyio==4.5.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.47.0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
youtube-search-python==1.6.6
requests==2.26.0
pymongo==4.9.1
pandas==2.2.3
beautifulsoup4==4.12.3
pycryptodome==3.18.0
aiohttp==3.9.1
google-api-python-client==2.95.0
motor==3.6.0
jinja2==3.1.2
PyJWT==2.8.0
bcrypt==4.1.2

# Flask and Quart Dependencies
quart==0.19.8   # Updated for compatibility
quart-cors==0.7.0
werkzeug==3.0.0

hypercorn
